requirements.txt
streamlit==1.53.1
tensorflow==2.20.0
numpy==2.4.1
pillow==12.1.0
scikit-learn==1.8.0
matplotlib==3.10.8
